Make MediaControllerImplBase local binder aware
The session->controller data flow already nicely avoids bundling if
we don't have to bundle by skipping the entire PlayerInfo object
which contains almost everything. But in the controller->session
data flow, some complex objects (MediaMetadata, MediaItem and
TrackSelectionParameters) were always bundled even if we don't have
to. This change skips bundling for these object and should improve
performance significantly when sending large lists of MediaItems to
a session in the same process.