Fix audio being disabled when `forceVideoTrack` is set to true
When `forceVideoTrack` is set to true to generate blank frames for gaps, it alters the order of `MediaSource` instances within the `MergingMediaSource`. This caused an issue in `CompositionTrackSelector`, which incorrectly disabled the main audio track.

`CompositionTrackSelector` is responsible for disabling the `SilenceMediaSource` track when the main media source contains its own audio (`shouldUseMediaAudio` is true). Previously, the selector assumed the silence track is the last in the list. However, when `forceVideoTrack` was enabled, the `mainMediaSource` was added after `SilenceMediaSource`, making the main audio the last track. As a result, the main audio was disabled by mistake.

This change corrects the issue by enforcing a consistent order where `SilenceMediaSource` is always the first source provided to `MergingMediaSource`. The `CompositionTrackSelector` is updated to use the fixed index `0` to correctly disable the silent audio track. Changes were also made to the index of the blank media source to accommodate the silence track changes.

PiperOrigin-RevId: 788926915