Don't call getCurrentTimeline when access is not granted
The current implementation of the session calls
`getCurrentTimeline()` when `COMMAND_GET_CURRENT_MEDIA_ITEM` is
granted but not `COMMAND_GET_TIMELINE`.

This change ensures we don't access the current item through the timeline. In such a case the `placeholder` flag of a timeline period is always set to false.

Issue: androidx/media#2665
PiperOrigin-RevId: 789229416