Use trackGroupIndex instead of assuming the index
Changed `CompositionTrackSelector` to no longer assume that the silence audio and blank image tracks are at index 0. It now uses the `trackGroupIndex` variable that was evaluated for each track group.

This removes the brittle assumption that required `SilenceMediaSource` and `BlankMediaSource` to be added before the `mainMediaSource` in `MergingMediaSource`.

The order of the `SilenceMediaSource` and `BlankMediaSource` in the list still needs to be maintained since we use their IDs to find their indices.

PiperOrigin-RevId: 789285123