Fix Compose test ahead of scheduling redesign
A drop of aosp/3208355 will result in a change of StandardTestDispatcher behavior that will clash with the current tests involving `withFrameMillis()` - special kind of delay that suspends the computation for a single Compose frame and is scheduled on the Compose frame clock directly. To avoid infiinfinitely scheduled work, we set `autoAdvance=false` (true by default in other tests).

PiperOrigin-RevId: 794600579