Remux `sample.mkv` to only have `Cue` entries for keyframes
This is recommended in the [Matroska
spec](https://www.matroska.org/technical/cues.html):
> It is RECOMMENDED to not reference non-keyframes of video tracks in
> Cues unless it references a Cluster Element which contains a CodecState
> Element but no keyframes.

The other files in the `mkv/` directory already follow this
recommendation.

This was done as follows:

```shell
$ mkvmerge -o sample_foo.mkv --cues 0:iframes sample.mkv
$ mv sample_foo.mkv sample.mkv
```

Unfortunately this re-muxes the file, which recalculates the sample
timestamps, hence the small diff in timestamps in the dump files. I
couldn't find a way to avoid this with `mkvmerge`.

PiperOrigin-RevId: 795013066