[ui-compose] Add ContentFrame - PlayerSurface with shutter and scaling
`ContentFrame` Composable will be the top level Composable for everything related to the display of media content. It aims to match the responsibility of the `AspectRatioFrameLayout` of `contentFrame` in the `PlayerView` world.

Current functionality includes covering with a shutter (black box by default) when the surface is not ready and resizing of the container according to the provided `ContentScale`.

In the future, this is the Composable that will host other subcomposables like artwork, image output, subtitles and error overlay.

PiperOrigin-RevId: 795452026