Use IdlingMediaCodecAdapterFactory in CapturingRenderersFactory
This ensures that the async codec operations are more predictable.

It also means we need to move CapturingRenderersFactory to Robolectric
utils, which should be fine given it's only used from Robolectric
anyway.

Also increase the interval at which IdlingMediaCodecAdapterFactory
idles its internal loopers to 5ms, which is sufficient because the
sole goal is to idle pending messages before the next doSomeWork call.
Increasing this interval makes the tests run much faster.

PiperOrigin-RevId: 796819925