Enable skipping of keyframe reset during scrubbing mode
If a forward seek is occurring within the same GoP then it is more efficient to continue decoding to the Seek position rather than flush the decoder and decoding previously processed frames.

This CL introduces new API versions of the `Renderer` methods `resetPosition` and `onPositionReset` to include a new parameter `sampleStreamIsResetToKeyFrame`. Custom renderers must upgrade to the new method as a result of this change.

PiperOrigin-RevId: 797344679