Refactor TestSimpleBasePlayer's constructor arg defaults
It is more logical for the empty TestSimpleBasePlayer to have no media items, but for the playlist to be empty, we also need an IDLE or ENDED playback state.

However, in most of the tests, we might need a 1-2 item playing player, so a util function provides a convenient overload.

PiperOrigin-RevId: 799982558