Add support for 13-channel audio layout
Updates `Util.getAudioTrackChannelConfig` to support a 13-channel configuration, utilizing new channel masks introduced in API level 32.

This is necessary to handle the output from certain on-device decoders, such as the Sony MPEG-H decoder for 360 Reality Audio, which outputs a 13-channel float PCM stream. Without this, attempts to configure the AudioTrack fail with an `Invalid output channel config`.

PiperOrigin-RevId: 800809455