support 20-bit pcm format constant
but don't support converting from/to 20-bit pcm. this is only useful
to communicate that an alac file is 20-bit, as a decoder is expected
to output 24-bit pcm instead (20-bit is not representable as integer
byte count, which causes many issues in the codebase if processing
would be attempted. hence, just don't, there's no reason to.)