Make the operations in PlayerIdFilteringAllocatorImpl atomic
The operation like `allocate()` or `release()` in `PlayerIdFilteringAllocatorImpl` has its internal logic pieces being locked by two locks separately. For example `release()`, the `allocator.release()` is locked by the `allocator` instance, and the `releaseInternal()` is locked by the `PlayerIdFilteringAllocatorImpl` instance. The same as `allocate()`. This will cause race condition issue when there are two threads A and B that one wants to `release()` and one wants to `allocate()`. Assume thread A calling `release(allocation)` gains the lock of internal allocator first, the internal allocator (`DefaultAllocator`) put the allocation back to the `availableAllocations` array, then waiting for the lock of `PlayerIdFilteringAllocatorImpl` to remove the allocation out of the map. Meanwhile thread B calling `allocate()` will gain the lock of internal allocator, which can take the allocation that just put back by A, then it waits for the lock of `PlayerIdFilteringAllocatorImpl` to add this allocation to the map. There is no guarantee that who will gain `PlayerIdFilteringAllocatorImpl` first. If B gains it first, then it will add the allocation to the map while the allocation has still not removed by A. The map doesn't record the "times" that the allocation being added, thus when A later gain that lock, it will remove the allocation from the map. And when B wants to release the allocation, there is no record in the map for it anymore, thus cause a `NullPointerException`.

PiperOrigin-RevId: 803501195