Control loading of PreloadMediaSource with LoadControl
We register a special `PlayerId.PRELOAD` to the `LoadControl` for `PreloadMediaSource`. In `DefaultLoadControl`, the mechanism of controlling loading based on the "PlayerId filtered allocation with the player-level target buffer bytes" can be applied.

The `LoadControl` now can be configured via `PreloadMediaSource.Factory` and plumbed to the created `PreloadMediaSource`. Also, an event `onLoadingUnableToContinue()` is added to `PreloadMediaSource.PreloadControl` for the caller being informed when the loading is unable to continue due to the `LoadControl` decision.

PiperOrigin-RevId: 805009716