Increase threshold for calling surface.setFrameRate
Every time this method is called, some internal screen refresh
vsync timing is reset too. That is, even if the screen refresh
rate is still the same, the value of vsyncTime % vsyncDuration
may jump around, causing an unexpectedly short or long gap
between two frames, which is noticeable for a user.

This means, we should be more cautious with setting a new frame
rate and only set it for meaningful changes. Changes of 0.1fps
are still quite small, but the higher limit avoids calling the
method after every few frames if there are some small variations
in the frame time deltas.

PiperOrigin-RevId: 805360781