Add hysteresis to vsync snapping
We snap the calculated release time to the closest vsync.
If there are any small variations and the release times
happen to be on the rounding boundary, we may easily introduce
smoothness issues by rounding in different directions, causing
frame freezes or drops.

This can be solved by some adding some hysteresis where we
keep rounding in the same direction in which we rounded before.
As many combinations of frame rate and screen refresh rate
oscillate between a clear match and being close the rounding
boundary (e.g. 60fps in 90Hz), we ignore the clear matches for
this logic completely.

PiperOrigin-RevId: 807810147