Remove format from UserDataCookie.
This change addresses a potential crash in Dav1dUserDataFreeCallback during decoder shutdown caused by attempts to call DeleteGlobalRef on the format member of the UserDataCookie, by removing the unused format field altogether.

The format passed from Java in DecoderInputBuffer is not read or used in the VideoDecoderOutputBuffer for AV1 decoding. Removing it from the UserDataCookie struct in dav1d_jni.cc eliminates the need to manage this JNI global reference.

PiperOrigin-RevId: 809085211