Fix automotive demo app by adding cast related fields to json
Cast requires a new field in the session demo json file. The parser
has no mercy and fails when that new attribute is not available, so
we need to add this field to each node in the json file for
automotive in the same way to avoid the app crashing on startup.

PiperOrigin-RevId: 813298843