Delay parsing of AV1 sequence headers until they are needed
Parsing the sequence header is costly, and is only needed if we're about to
skip or drop an input buffer.
When a key frame sample is queued to the AV1 sample dependency parser,
copy a prefix of the data into a temporary buffer, and only parse it if we're
able to skip or drop a frame.

PiperOrigin-RevId: 814246949