Revert `compileSdk` to 35 and hard-code some constants in tests for now
Increasing the `compileSdk` forces all apps depending on the library to
increase their `compileSdk` too, which can force them to upgrade their
AGP and Gradle versions too:
https://developer.android.com/build/releases/gradle-plugin#api-level-support

Forcing this only so we can avoid hard-coding some constants in tests
doesn't seem worth it. We will bump back to 36 when we have a stronger
need to depend on something only available at that API level.

PiperOrigin-RevId: 815711168