[ui-compose] Fix lint errors about Modifier factory
Used to be a warning, but now became an error: "Modifier factory functions must use the receiver Modifier instance".

By starting with this.then(...), we are explicitly applying the new modifiers to the receiver Modifier instance, which is the standard way to build up a chain of modifiers in Compose. While the previous code might have compiled, it wasn't following the recommended pattern for Modifier factory function.

PiperOrigin-RevId: 817123087