Ensure errors during simulated manifest refresh do not crash app
The message to simulate manifest refreshs is run on a regular
Handler message without the try/catch logic from ExoPlayer. The
method may throw for malformed media though, so we need to make
sure to report the error via the usual error reporting chain
instead of directly crashing the app.

Issue: androidx/media#2805
PiperOrigin-RevId: 817669651