Replace `forceAudio/VideoTrack` with `trackTypes`
This change introduces a `Set<@C.TrackType Integer> trackTypes` to `EditedMediaItemSequence` to explicitly define which track types are expected in the sequence's output. The deprecated `experimentalSetForceAudioTrack` and `experimentalSetForceVideoTrack` methods now modify this set. Asset loaders and `CompositionPlayer` are updated to respect these `trackTypes`.

The rest of the tests will be migrated in followup CLs

PiperOrigin-RevId: 818574408