Add CommandButton.executeAction to allow controllers to trigger action
This makes it possible to easily trigger the intended action from
the CommandButton instance. To support actions requiring a parameter,
this adds another override to CommandButton.Builder.

The current change does not modify which buttons are shared with the
platform session and what older Media3 controllers can support. It
only enables new Media3 controllers to trigger the actions where
possible.

PiperOrigin-RevId: 823019114