Ensure extras for custom commands/button are always forwarded
Media3 has two different extras fields, one in SessionCommand
and one in CommandButton that can be used for slightly different
purposes. In the platform session's CustomAction, only one args
Bundle is available.

The current code has very opionated mapping how we convert the
two extras into one, and the one extra into two. This causes
some inconsistencies in the received information when a controller
triggers a custom command, depending on whether the other end
of the communication uses platform or Media3 code.

This can be made more consistent and less opionated by always
keeping all extras available. When going from two Bundles to
one, we can merge them together. From one Bundle to two, we
can use the same Bundle in both places. This may duplicate
some information that was not initially meant to be set, but
receiver of Bundle data should always check against their
expected keys only, so that having additional keys is not an
issue.

PiperOrigin-RevId: 823071333