Fix issue where scrubbing near end could seek to next media item
Prevent skipping keyframe reset when renderers have had their decoders flushed while processing a seek. This can occur when seeking near the end of a media item while playing a playlist or with repeat mode. This ensures the player correctly handles the seek within the current item's bounds.

The solution is that a renderer that has not received a keyframe yet must return false to `supportsResetPositionWithoutKeyFrameReset`.

PiperOrigin-RevId: 835139084